requirements.txt
streamlit
crewai
openai
load-dotenv